+++ scripts/ArrowStick/global.lua
@@ -1,8 +1,7 @@
-local _, world = pcall(require, "openmw.world")
-local _, async = pcall(require, "openmw.async")
-local util = require("openmw.util")
+local world = require("openmw.world")
 local storage = require("openmw.storage")
 local I = require("openmw.interfaces")
+local core = require("openmw.core")
 
 local IE = require("scripts.ArrowStick.utils.impactEffects")
 local consts = require("scripts.ArrowStick.utils.consts")
@@ -12,24 +11,8 @@ local settingsImpactEffects = storage.globalSection("SettingsArrowStick_impactEf
 local arrowDespawnScript = "scripts/ArrowStick/customArrow.lua"
 
 local shotArrows = {}
-local xrot
-local xpos
 
-local function rotateArrow(data)
-    local obj = data.obj
-    data.obj:teleport(obj.cell, obj.position, data.rotation)
-end
-
-local function onItemActive(item)
-    if xrot and xpos then
-        async:newUnsavableSimulationTimer(0.1, function()
-            item:teleport(item.cell.name, xpos, xrot)
-            xrot = nil
-        end)
-    end
-end
-
-local function placeArrow(data)
+local function placeNewArrow(data)
     local id = data.id
     local pos = data.position
     local rot = data.rotation
@@ -41,28 +24,33 @@ local function placeArrow(data)
         local mat = IE.getMaterial(data.hitObj, hitWater)
 
         if settingsImpactEffects:get("impactEffects") then
-            local vfxPos = hitWater and waterPos or pos
-            IE.addImpactEffects(data.weapon, vfxPos, mat, player.position)
+            I.impactEffects.spawnEffect({
+                material = mat,
+                hitPos = hitWater and waterPos or pos,
+            })
         end
 
-        if settingsImpactEffects:get("checkMaterial") then
-            if consts.unstickableMaterials[mat] then return end
+        if settingsImpactEffects:get("checkMaterial")
+            and consts.unstickableMaterials[mat]
+        then
+            return
         end
     end
 
     if hitWater and not settings:get("stickUnderwater") then return end
 
-    local temppos = util.vector3(pos.x, pos.y, pos.z - 1000)
     local newArrow = world.createObject(id)
-    newArrow:teleport(player.cell.name, temppos, rot)
-
-    xrot = rot
-    xpos = util.vector3(pos.x, pos.y, pos.z)
+    newArrow:teleport(player.cell.name, pos, rot)
 
     if settings:get("despawnArrows") then
         newArrow:addScript(arrowDespawnScript)
         shotArrows[newArrow.id] = newArrow
     end
+
+    core.sendGlobalEvent("ArrowStick_ArrowPlaced", {
+        item = newArrow,
+        position = pos,
+    })
 end
 
 local function onSave()
@@ -91,14 +79,12 @@ end
 
 return {
     engineHandlers = {
-        onItemActive = onItemActive,
         onActivate = onActivate,
         onSave = onSave,
         onLoad = onLoad,
     },
     eventHandlers = {
-        rotateArrow = rotateArrow,
-        placeArrow = placeArrow,
-        arrowInactive = arrowInactive,
+        ArrowStick_PlaceNewArrow = placeNewArrow,
+        ArrowStick_ArrowInactive = arrowInactive,
     }
 }